fix: moved store to window into app.vue
